rt haskell: add Reader monad to Haskell
This uses the Reader monad to define the function called in the
while condition of `awaken`. This will have some proof impact,
esp for getObject, although it should be possible to restore the
old definition of getObject via proof.

I'm not fully satisfied with this solution yet, because the type
of the Reader monad is different to what we use in Isabelle. It'd
be nice for these to coincide, although that may not make sense
in Haskell.

Signed-off-by: Gerwin Klein <gerwin.klein@data61.csiro.au>
Signed-off-by: Miki Tanaka <miki.tanaka@data61.csiro.au>